% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judge_Pot.R
\name{judge_Pot}
\alias{judge_Pot}
\title{Judge the potential}
\usage{
judge_Pot(
  Curr_acd,
  predict_acd,
  pot_lvl1 = 15,
  pot_lvl2 = 10,
  abs_lvl1 = 20,
  abs_lvl2 = 10
)
}
\arguments{
\item{Curr_acd}{Numerical, current academic performance, standardized.}

\item{predict_acd, }{Numerical, predicted academic performance.}

\item{pot_lvl1, pot_lvl2}{Numerical, the amount of potential needed for different level}

\item{abs_lvl1, abs_lvl2}{Numerical, the absolute level away from average level(100) needed
for different level}
}
\value{
A factor vector, in which the factor levels are the predefined judgment.
}
\description{
Judge the potential of a student by his/her current academic performance
and predicted academic performance. It is assumed that the difference between
current academic performance and predicted academic performance is potential on academic.
}
