% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Quik Histogram Plot}
\usage{
plot_distribution(data, x = game_score_raw, group = NULL, bins = 50)
}
\arguments{
\item{data}{Data.frame, the data}

\item{x}{Unquoted expression, colname of the observation value, default game_score_raw}

\item{group}{Formula, indicate the facet variables}

\item{bins}{Numerical, how much bins are used in geom_histogram, default 50}
}
\value{
a ggplot object
}
\description{
This function is a wrapper of \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}, usually used to
check the distribution of game scores or academic scores. If group is assigned, it will
plot the distribution by groups (usually game_name or subject_name).
}
