% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pred.R
\name{plot_pred}
\alias{plot_pred}
\title{Plot the prediction and judgment}
\usage{
plot_pred(model_report)
}
\arguments{
\item{model_report}{List, the output of function \code{\link[=prepare_report]{prepare_report()}}.}
}
\value{
A ggplot object
}
\description{
This function is used to visualize the result of prediction and
judgment.
}
