% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{plot the scatter plot of game score and academic score by group}
\usage{
plot_scatter(
  data,
  ob_value_x = "game_score_raw",
  ob_value_y = "acd_score",
  group,
  cor = FALSE
)
}
\arguments{
\item{data}{Data.frame, contains two observation value column and groups column}

\item{ob_value_x, ob_value_y}{String, colname of the observation value}

\item{group}{Formula, indicate the facet variables}

\item{cor}{Logical, indicate whether plot correlation or not}
}
\value{
a ggplot object
}
\description{
This function is used to check the scatter distribution of
game scores and academic scores, correlation will be calculate if specified
}
