% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_report.R
\name{prepare_report}
\alias{prepare_report}
\title{Summary all analysis and generate essential parts for final model report}
\usage{
prepare_report(
  data_wider,
  model,
  abilities,
  class_NL,
  background = NULL,
  plot_x = "Z_acd_score",
  plot_y = "wt_sum_predict"
)
}
\arguments{
\item{data_wider}{Data.frame, with each row represent all observation of a person
and his/her missing_rate, prediction. ie. the output of function wt_sum_predict}

\item{model}{A lm object, with additional rw part, generated by select_by_RW.}

\item{abilities}{Data.frame, including two column, the game_name and the represent
ability of it.}

\item{class_NL}{Data.frame, class name list, has a column indicate the name and
a corresponding column indicate it's class. user_name and class}

\item{background}{Data.frame or NULL, if the background is specified, it will be
used in function \code{\link[=plot_pred]{plot_pred()}} to visualize the judgment standard. It should have
x, y and judgment as columns.}

\item{plot_x, plot_y}{String, the x axis and y axis variables used in function \code{\link[=plot_pred]{plot_pred()}}}
}
\value{
A list, contains following values:
\item{report_rw}{The tasks in model and its corresponding ability and relative weight}
\item{n_task}{The number of tasks in model}
\item{r2,adj_r2}{Model interpretation rate}
\item{summary_pred}{The judgment and it's description and people proportion}
\item{detail_pred}{The details of prediction and judgment, for every person}
\item{namelist_pred}{The prediction and judgment name list for every class}
\item{namelist_drop}{The name list of people those who are not involved prediction}
\item{background}{The background used in \code{\link[=plot_pred]{plot_pred()}}}
\item{plot_x,plot_y}{The x axis and y axis variables used in function \code{\link[=plot_pred]{plot_pred()}}}
}
\description{
After former pipeline, there are a lost of results, this function
is used to tidy and summary the results for final model report.
}
