% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_GameFinish.R
\name{rate_GameFinish}
\alias{rate_GameFinish}
\title{Calculate the finish rate of every game and plot it}
\usage{
rate_GameFinish(
  data,
  plot = FALSE,
  idx_game = "game_name",
  idx_user = "user_id",
  ob_value = "game_score_raw"
)
}
\arguments{
\item{data}{Data.frame, contains the essential data(index of game, index of user,
observed value).}

\item{plot}{Logical value indicate if plot the finish rate or not, default \code{FALSE}}

\item{idx_game}{String, colname of the unique index of game, default "game_name"}

\item{idx_user}{String, colname of the unique index of user, default "user_id"}

\item{ob_value}{String, colname of the observed value, default "game_score_raw"}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{idx_game}{index of game, will be the same as input.}
\item{finish_rate}{finish rate of the games}
}
\description{
This function is used to check the finish rate of multiple games,
since there is always some missing value in the iquizoo data. To get a more
complete data, it is useful to drop some games that finish rate is too low.
Given a \code{data.frame} contains essential columns (usually "game_name",
"user_name"/"user_id", and "game_score_raw"), this function will calculate
the finish rate of every game and plot it if needed.
}
