% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPT_model_tasks.R
\name{RPT_model_tasks}
\alias{RPT_model_tasks}
\title{Report model tasks}
\usage{
RPT_model_tasks(model, abilities)
}
\arguments{
\item{model}{A lm object, with additional rw part, generated by select_by_RW.}

\item{abilities}{Data.frame, including two essential column, the game_name and the represent
ability of it(ab_name_second). And a alternative column, ab_type.}
}
\value{
A \link[tibble:tibble-package]{tibble} can be used to report directly
}
\description{
This function summary the result of linear model and relative
weight analysis, and then make a tibble for report
}
