% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPT_namelist_drop.R
\name{RPT_namelist_drop}
\alias{RPT_namelist_drop}
\title{The namelist that we don't make prediction}
\usage{
RPT_namelist_drop(data_wider, class_NL, threshold = 30)
}
\arguments{
\item{data_wider}{Data.frame, which each row contains all the observations of a single person}

\item{threshold}{Numerical, missing proportion threshold above which the user will be dropped}
}
\value{
A \link[tibble:tibble-package]{tibble}/table can be used to report directly
}
\description{
This function is used to make name list
that we didn't make prediction, on which users had
too many predictors missing
}
