% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPT_namelist_pred.R
\name{RPT_namelist_pred}
\alias{RPT_namelist_pred}
\title{Make a prediction name list of every class}
\usage{
RPT_namelist_pred(class_NL, user_NL)
}
\arguments{
\item{class_NL}{Data.frame, class name list, has a column indicate the name and
a corresponding column indicate it's class. user_name and class}

\item{user_NL}{Data.frame or character vector, if input is a data.frame, it should has a column
indicate the name and a corresponding potential judgment(see \code{\link[=judge_Pot]{judge_Pot()}} for more information).
If input is a character vector, it will be considered as user name.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains class and the name list of every class as a single long string
}
\description{
To present the prediction details and those people who are not
involved in prediction for some reason, it is a good idea to make a name list
for every class. So the teachers can easily know the details in his/her class.
}
