% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPT_plot_pred.R
\name{RPT_plot_pred}
\alias{RPT_plot_pred}
\title{Plot the prediction and judgment}
\usage{
RPT_plot_pred(
  data_wider,
  x = Z_acd_score,
  y = wt_sum_predict,
  color = judgment,
  r2 = RPT_r2(Curr_mod, adj = TRUE)
)
}
\arguments{
\item{data_wider}{Data.frame, which each row contains all the observations of a single person.}

\item{x, y}{Unquoted expression, colname of the observation value}

\item{color}{Unquoted expression, colname of the potential judgment, default judgment}
}
\value{
A ggplot object
}
\description{
This function is used to visualize the result of prediction and
judgment.
}
