% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPT_summary_pred.R
\name{RPT_summary_pred}
\alias{RPT_summary_pred}
\title{Summary prediction and potential judgment}
\usage{
RPT_summary_pred(data_wider, threshold = 30)
}
\arguments{
\item{data_wider}{Data.frame, which each row contains all the observations of a single person}

\item{threshold}{Numerical, missing proportion threshold above which the user will be dropped}
}
\value{
A \link[tibble:tibble-package]{tibble}/table can be used to report directly
}
\description{
This function is used to make a summary table of academic potential
prediction and potential judgment.
}
