% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select_by_RW.R
\name{select_by_RW}
\alias{select_by_RW}
\title{Select predictor games by relative weight of predictors}
\usage{
select_by_RW(data_wider, formula, n_task = 5, least_RW = NULL)
}
\arguments{
\item{data_wider}{Data.frame, which each row contains all the observations of a single person.}

\item{formula}{Formula, indicate which variables should be included in model. Or a lm object,
the formula will be the call of model, ie. model$call.}

\item{n_task}{Numerical, a selection standard indicate how many tasks preserve.}

\item{least_RW}{Numerical, a selection standard indicate the least relative weight
the final selected task should have.}
}
\value{
a lm object with relative weight, "rw"
}
\description{
Usually users will have lots of iquizoo games test at once,
however it is bad idea to put all the games into a model to predict academic
score, which might cause the problem of overfitting or a decline of adjust r squared.
So we choose a subset of iquizoo games to be placed into the model by drop the
least import variable, according to relative analysis.
}
