% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_sum_predict.R
\name{wt_sum_predict}
\alias{wt_sum_predict}
\title{Predict by weighted sum}
\usage{
wt_sum_predict(data_wider, model)
}
\arguments{
\item{data_wider}{Data.frame, which each row contains all the observations of a single person.}

\item{model}{A lm object, the fitted linear model.}
}
\value{
A vector with length nrow(data_wider), indicate the weighted sum prediction
}
\description{
This is a attempt to make prediction to observation with missing
predictors using the weighted sum from a linear model
}
