% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_missing_prop.R
\name{calc_missing_prop}
\alias{calc_missing_prop}
\title{Calculate the missing proportion of predictors by relative weight}
\usage{
calc_missing_prop(data_wider, RelativeWeight)
}
\arguments{
\item{data_wider}{Data.frame, which each row contains all the observations of a single person.}

\item{RelativeWeight}{Data.frame, generate by rwa::rwa, contains the relative weight of
each variable}
}
\value{
A \link[tibble:tibble-package]{tibble} with the missing proportion of predictors add on the right
}
\description{
There is a situation that we want to make prediction for
incomplete observation, which a linear model don't support. However, we try to
approximitly predict those imcomplete observation with small missing part.
So, this function is used to calculate the missing proportion of predictors,
the weight fo each predictor is from relative weight analysis.
}
