% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot the distribution by group}
\usage{
plot_distribution(
  data,
  ob_value = "game_score_raw",
  group = ~game_name,
  bins = 50
)
}
\arguments{
\item{data}{Data.frame, contains group column and observation value column}

\item{ob_value}{String, colname of the observed value, default "game_score_raw"}

\item{group}{Formula, indicate the facet variables, default ~ game_name}

\item{bins}{Numerical, how much bins are used in geom_histogram, default 50}
}
\value{
a ggplot object
}
\description{
Check the distribution of game scores or academic scores is often
needed, this function is used to do this. It plot the distribution by groups
(game_name or subject_name).
}
