% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Quick Scatter Plot}
\usage{
plot_scatter(
  data,
  x = game_score_raw,
  y = acd_score,
  group = NULL,
  cor = FALSE
)
}
\arguments{
\item{data}{Data.frame, the data}

\item{x, y}{Unquoted expression, colname of the observation value}

\item{group}{Formula, indicate the facet variables}

\item{cor}{Logical, indicate whether plot correlation or not}
}
\value{
a ggplot object
}
\description{
This function is a wrapper of \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}, usually
used to check the scatter distribution of game scores and academic scores,
correlation will be calculate if specified
}
