% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_template.R
\name{prepare_template}
\alias{prepare_template}
\title{prepare the prediction template}
\usage{
prepare_template()
}
\arguments{
\item{None}{}
}
\value{
None
}
\description{
copy the prediction template into current work directory
}
